fix!: Lookup g1/g2 measures, test proc, test data by criterion ID
